Address @ghostwords comments and combine canvas suites
